clean: reorganize helper files, move them out of jest mocks directory
